relocate Bodu.Text.Formats-only contracts and KATs out of Bodu.Test
Three contract bases and four KAT records were consumed only by
Bodu.Text.Formats.Test, so they move alongside their consumers per the
single-consumer rule.

Moved to Bodu.Text.Formats/test/Contracts/ (namespace
Bodu.Text.Formats.Contracts — a new shared namespace, since concrete
subclasses are spread across four per-format namespaces:
Bodu.Text.{Bencode,Delimited,DotEnv,Ini}.Contracts):

Contracts:
- BinaryDocumentFormatContractTests<TDocument, TOptions>
- TextDocumentFormatContractTests<TDocument, TOptions>
- StreamRoundTripContractTests<T>

KATs:
- BinaryDocumentKat<TDocument, TOptions>
- InvalidBinaryDocumentKat<TOptions>
- TextDocumentKat<TDocument, TOptions>
- InvalidTextDocumentKat<TOptions>

Subclass `using` updates (across 6 contract subclass files):
- Replaced `using Bodu.Test.Contracts;` with
  `using Bodu.Text.Formats.Contracts;`.
- Removed `using Bodu.Test.Kat;` from format-document subclasses (KATs
  are now in Bodu.Text.Formats.Contracts).
- Retained `using Bodu.Test.Kat;` on BencodeStreamRoundTripContractTests
  because it still references the generic BinaryKat<T, T> which stays
  in Bodu.Test.

Domain-local KAT records inspected but NOT consolidated (different
abstractions — richer fields like Source citations and per-format
structured entries):
- BencodeKnownAnswerVector
- DelimitedKnownAnswerVector
- DotEnvKnownAnswerVector
- IniKnownAnswerVector

Verified: dotnet build passes 0 errors;
dotnet test --settings bvt.runsettings passes 636/636.
