+++ Bodu.Text.Formats/test/Contracts/BinaryDocumentKat.cs
@@ -1,10 +1,12 @@
-﻿// ---------------------------------------------------------------------------------------------------------------
+// ---------------------------------------------------------------------------------------------------------------
 // <copyright file="BinaryDocumentKat.cs" company="Bodu Pty. Ltd.">
 //     Copyright (c) Bodu Pty. Ltd.. All rights reserved.
 // </copyright>
 // ---------------------------------------------------------------------------------------------------------------
 
-namespace Bodu.Test.Kat;
+using Bodu.Test.Kat;
+
+namespace Bodu.Text.Formats.Contracts;
 
 /// <summary>
 /// Represents a known-answer test row for a decode/encode binary-document format such as Bencode. A row pairs the raw
