+++ Bodu.Text.Formats/test/Contracts/InvalidBinaryDocumentKat.cs
@@ -1,10 +1,12 @@
-﻿// ---------------------------------------------------------------------------------------------------------------
+// ---------------------------------------------------------------------------------------------------------------
 // <copyright file="InvalidBinaryDocumentKat.cs" company="Bodu Pty. Ltd.">
 //     Copyright (c) Bodu Pty. Ltd.. All rights reserved.
 // </copyright>
 // ---------------------------------------------------------------------------------------------------------------
 
-namespace Bodu.Test.Kat;
+using Bodu.Test.Kat;
+
+namespace Bodu.Text.Formats.Contracts;
 
 /// <summary>
 /// Represents a known-failure row for a binary-document decoder. The decoder must reject <see cref="Payload" /> by
